Before deleting an orphan, ask the object if it should really be deleted using preventOrphanDeletion.
I want to be able to distinguish between new content on the device that the server doesn't know about and
objects that have been deleted on the server that I need to expunge.  preventOrphanDeletion gives me a
mechanism for doing this.  When an object is being prepared for posting in a background operation I want
to make sure that it is never deleted by the RestKit orphan deletion system.